pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "arey"
version = "0.0.6"
requires-python = '>=3.10'
description = 'Simple large language model playground.'
authors = [{ name = 'codito', email = 'codito@codito.in' }]
dependencies = [
    "click>=8.1.7",
    "markdown>=3.5.2",
    "ollama>=0.1.6",
    "openai>=1.12.0",
    "pydantic>=2.9.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "tiktoken>=0.6.0",
    "watchfiles>=0.21.0",
    "wurlitzer>=3.0.3",
]
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Environment :: Console',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Intended Audience :: Developers',
    'Intended Audience :: Information Technology',
    "License :: OSI Approved :: MIT License",
    'Operating System :: POSIX :: Linux',
    'Operating System :: Microsoft :: Windows',
    'Operating System :: MacOS',
    'Topic :: Communications :: Chat',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
    'Topic :: Terminals',
    'Topic :: Utilities',
    'Topic :: Text Processing :: Markup :: Markdown',
]
readme = { file = "README.md", content-type = "text/markdown" }

[project.optional-dependencies]
llama = ["llama-cpp-python>=0.3.1"]
samples = ["beautifulsoup4", "lxml"]

[project.scripts]
arey = "arey.main:main"

[project.urls]
Homepage = 'https://github.com/codito/arey'
Documentation = 'https://apps.codito.in/arey'
Source = 'https://github.com/codito/arey'
Changelog = 'https://github.com/codito/arey/blob/master/CHANGELOG.md'

[tool.uv]
dev-dependencies = [
    "basedpyright",
    "mkdocs",
    "mkdocs-material",
    "pre-commit",
    "pyfakefs",
    "pytest",
    "pytest-mock",
    "pytest-cov",
    "ruff",
]

[tool.setuptools]
include-package-data = false

[tool.setuptools.package-data]
"arey" = ["**/*.yml", "**/*.md"]

[tool.setuptools.packages.find]
where = ["."]       # ["."] by default
include = ["arey*"] # ["*"] by default
exclude = []        # empty by default
namespaces = false  # true by default

[tool.coverage.run]
omit = []

[tool.coverage.report]
show_missing = "true"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=arey --cov-report=term-missing --cov-report=json --quiet"

[tool.ruff]
extend-exclude = [".venv", "venv"]
line-length = 88

[tool.ruff.lint]
ignore = []
select = ["D", "E", "F", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D101", "D102", "D103"]
"**/*.py" = ["F405", "D203", "D213"]

[tool.basedpyright]
reportUnusedCallResult = "none"
